+++ dc-cudami-client/src/main/java/de/digitalcollections/cudami/client/identifiable/entity/CudamiArticlesClient.java
@@ -4,7 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import de.digitalcollections.model.exception.TechnicalException;
 import de.digitalcollections.model.exception.http.client.ResourceNotFoundException;
 import de.digitalcollections.model.identifiable.entity.Article;
+import de.digitalcollections.model.identifiable.entity.agent.Agent;
 import java.net.http.HttpClient;
+import java.util.List;
 import java.util.Locale;
 import java.util.UUID;
 
@@ -24,5 +26,24 @@ public class CudamiArticlesClient extends CudamiEntitiesClient<Article> {
     } catch (ResourceNotFoundException e) {
       return null;
     }
+  }
+
+  public boolean removeCreator(UUID articleUuid, UUID agentUuid) throws TechnicalException {
+    try {
+      doDeleteRequestForString(
+          String.format(baseEndpoint + "/%s/creators/%s", articleUuid, agentUuid));
+    } catch (ResourceNotFoundException e) {
+      return false;
+    }
+    return true;
+  }
+
+  public boolean addCreators(UUID articleUuid, List<Agent> agents) throws TechnicalException {
+    try {
+      doPostRequestForString(String.format(baseEndpoint + "/%s/creators", articleUuid), agents);
+    } catch (ResourceNotFoundException e) {
+      return false;
+    }
+    return true;
   }
 }